requirements.txt
streamlit
pandas
openpyxl
langchain
openai
python-dotenv
fpdf
matplotlib
langchain-community
langchain_experimental
tabulate
langchain-openai